Intermediate version

+++ System/process/ObjectTrackAct.cxx
@@ -120,7 +120,8 @@ ObjectTrackAct::clearAll()
 }
 
 void
-ObjectTrackAct::addUnit(const Simulation& System,const int objN,
+ObjectTrackAct::addUnit(const Simulation& System,
+			const long int objN,
 			const Geometry::Vec3D& IPt)
   /*!
     Create a target track between the IPt and the target point
@@ -132,18 +133,18 @@ ObjectTrackAct::addUnit(const Simulation& System,const int objN,
   ELog::RegMethod RegA("ObjectTrackAct","addUnit");
 
   // Remove old track
-  std::map<int,LineTrack>::iterator mc=Items.find(objN);
+  std::map<long int,LineTrack>::iterator mc=Items.find(objN);
   if (mc!=Items.end())
     Items.erase(mc);
 
   LineTrack A(IPt,TargetPt);
   A.calculate(System);
-  Items.insert(std::map<int,LineTrack>::value_type(objN,A));
+  Items.insert(std::map<long int,LineTrack>::value_type(objN,A));
   return;
 }  
 
 double
-ObjectTrackAct::getMatSum(const int objN) const
+ObjectTrackAct::getMatSum(const long int objN) const
   /*!
     Calculate the sum in the material
     \param objN :: Cell number to use
@@ -152,9 +153,9 @@ ObjectTrackAct::getMatSum(const int objN) const
 {
   ELog::RegMethod RegA("ObjectTrackAct","getMatSum");
 
-  std::map<int,LineTrack>::const_iterator mc=Items.find(objN);
+  std::map<long int,LineTrack>::const_iterator mc=Items.find(objN);
   if (mc==Items.end())
-    throw ColErr::InContainerError<int>(objN,"objN in Items");
+    throw ColErr::InContainerError<long int>(objN,"objN in Items");
   
   // Get Two Paired Vectors
   const std::vector<MonteCarlo::Object*>& OVec=
@@ -169,7 +170,7 @@ ObjectTrackAct::getMatSum(const int objN) const
 }
 
 double
-ObjectTrackAct::getAttnSum(const int objN) const
+ObjectTrackAct::getAttnSum(const long int objN) const
   /*!
     Calculate the sum in the material
     \param objN :: Cell number to use
@@ -181,9 +182,9 @@ ObjectTrackAct::getAttnSum(const int objN) const
   const ModelSupport::DBMaterial& DB=
     ModelSupport::DBMaterial::Instance();
 
-  std::map<int,LineTrack>::const_iterator mc=Items.find(objN);
+  std::map<long int,LineTrack>::const_iterator mc=Items.find(objN);
   if (mc==Items.end())
-    throw ColErr::InContainerError<int>(objN,"objN in Items");
+    throw ColErr::InContainerError<long int>(objN,"objN in Items");
   
   // Get Two Paired Vectors
   const std::vector<MonteCarlo::Object*>& OVec=
@@ -194,10 +195,11 @@ ObjectTrackAct::getAttnSum(const int objN) const
   double sum(0.0);
   for(size_t i=0;i<TVec.size();i++)
     {
-      const int matN=OVec[i]->getMat();
+      const long int matN=OVec[i]->getMat();
       if (matN)
 	{
-	  const MonteCarlo::Material& matInfo=DB.getMaterial(matN);
+	  const MonteCarlo::Material& matInfo=
+	    DB.getMaterial(static_cast<int>(matN));
 	  const double density=matInfo.getAtomDensity();
 	  const double AMean=matInfo.getMeanA();
 	  sum+=TVec[i]*std::pow(AMean,0.66)*density;
@@ -207,7 +209,7 @@ ObjectTrackAct::getAttnSum(const int objN) const
 }
 
 double
-ObjectTrackAct::getDistance(const int objN) const
+ObjectTrackAct::getDistance(const long int objN) const
   /*!
     Calculate the sum of the distance
     \param objN :: Cell number to use
@@ -216,14 +218,14 @@ ObjectTrackAct::getDistance(const int objN) const
 {
   ELog::RegMethod RegA("ObjectTrackAct","getDistance");
 
-  std::map<int,LineTrack>::const_iterator mc=Items.find(objN);
+  std::map<long int,LineTrack>::const_iterator mc=Items.find(objN);
   if (mc==Items.end())
-    throw ColErr::InContainerError<int>(objN,"objN in Items");
+    throw ColErr::InContainerError<long int>(objN,"objN in Items");
   return mc->second.getTotalDist();
 }
 
 void
-ObjectTrackAct::createAttenPath(std::vector<int>& cellN,
+ObjectTrackAct::createAttenPath(std::vector<long int>& cellN,
 				std::vector<double>& attnD) const
   /*!
     Calculate the sum of the distance
@@ -231,7 +233,7 @@ ObjectTrackAct::createAttenPath(std::vector<int>& cellN,
   */
 {
   ELog::RegMethod RegA("ObjectTrackAct","createAttenPath");
-  for(const std::map<int,LineTrack>::value_type& mc : Items)
+  for(const std::map<long int,LineTrack>::value_type& mc : Items)
     mc.second.createAttenPath(cellN,attnD);
   return;
 }
@@ -246,7 +248,7 @@ ObjectTrackAct::write(std::ostream& OX) const
   ELog::RegMethod RegA("ObjectTrackAct","write");
 
   OX<<"WRITE"<<std::endl;
-  std::map<int,LineTrack>::const_iterator mc;
+  std::map<long int,LineTrack>::const_iterator mc;
   for(mc=Items.begin();mc!=Items.end();mc++)
     OX<<mc->first<<" : "<<mc->second<<std::endl;
   return;
